Частичные правки (не справился с 1 правкой)
1. Сделал отдельные модули для функций js.
2. Исправил email и addProduct как модификатор элемента modal__container
3. сделал цвета через hex
4. добавил БЭМ-блок для кнопки
