+++ source/index.html
@@ -433,22 +433,22 @@
     
     <div class="modal">
       
-        <div class="modal__container modal__add-product ">
-          <button type="button" class="modal__button-close">
-            <img src="img/close_icon.svg" alt="Закрыть меню">
-          </button>
-          <h2 class="modal__title">Товар успешно<br>добавлен в корзину</h2>
-          <button type="button" class="modal__button">Продолжить покупки</button>
-      </div>
-      
-      <div class="modal__container modal__email">
+      <div class="modal__container modal__container--add-product">
         <button type="button" class="modal__button-close">
           <img src="img/close_icon.svg" alt="Закрыть меню">
         </button>
-        <h2 class="modal__title modal__email-title">Проверьте ваш email и подтвердите подписку</h2>
-      </div>
-      
+        <h2 class="modal__title">Товар успешно<br>добавлен в корзину</h2>
+        <button type="button" class="button modal__button">Продолжить покупки</button>
     </div>
-    <script src="js/main.js" defer type="module"></script>
-  </body>
+    
+    <div class="modal__container modal__container--email">
+      <button type="button" class="modal__button-close">
+        <img src="img/close_icon.svg" alt="Закрыть меню">
+      </button>
+      <h2 class="modal__title modal__email-title">Проверьте ваш email и подтвердите подписку</h2>
+    </div>
+    
+  </div>
+  <script src="js/main.js" defer type="module"></script>
+</body>
 </html>
