+++ source/js/main.js
@@ -1,6 +1,9 @@
-import './script.js';
+import './addProduct.js';
+import './openMenu.js';
+import './modalEmail.js';
 import renderProducts from './renderProducts.js';
 import products from './products.js';
+
 
 const catalogList = document.querySelector('.best-selling__product-wrapper');
 const catalogItemTemplate = document.querySelector('#product').content;
