+++ source/sass/style.scss
@@ -5,8 +5,8 @@
 
 @import 'root.scss';
 @import 'blocks/container.scss';
-@import 'blocks/modal.scss';
 @import 'blocks/button.scss';
+@import 'blocks/modal.scss';
 @import 'blocks/show-all.scss';
 @import 'blocks/banner.scss';
 @import 'blocks/slider-pagination.scss';
